fix(docs): address Copilot review on conformance harness (#1148)
